+++ dflib/src/test/java/com/nhl/dflib/print/TabularPrinter_DataFrameTest.java
@@ -2,6 +2,8 @@ package com.nhl.dflib.print;
 
 import com.nhl.dflib.DataFrame;
 import com.nhl.dflib.Index;
+import com.nhl.dflib.IntSeries;
+import com.nhl.dflib.Series;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -14,11 +16,9 @@ public class TabularPrinter_DataFrameTest {
     @Before
     public void initDataFrame() {
         Index columns = Index.forLabels("col1", "column2");
-        this.df = DataFrame.forSequenceFoldByRow(columns,
-                "one", 1,
-                "two", 2,
-                "three", 3,
-                "four", 4);
+        this.df = DataFrame.forColumns(columns,
+                Series.forData("one", "two", "three", "four"),
+                IntSeries.forInts(1, 2, 3, 44));
     }
 
     @Test
@@ -28,10 +28,10 @@ public class TabularPrinter_DataFrameTest {
         assertEquals(System.lineSeparator() +
                 "col1  column2" + System.lineSeparator() +
                 "----- -------" + System.lineSeparator() +
-                "one   1      " + System.lineSeparator() +
-                "two   2      " + System.lineSeparator() +
-                "three 3      " + System.lineSeparator() +
-                "four  4      " + System.lineSeparator() +
+                "one         1" + System.lineSeparator() +
+                "two         2" + System.lineSeparator() +
+                "three       3" + System.lineSeparator() +
+                "four       44" + System.lineSeparator() +
                 "4 rows x 2 columns", p.toString(df));
     }
 
@@ -42,8 +42,8 @@ public class TabularPrinter_DataFrameTest {
         assertEquals(System.lineSeparator() +
                 "col1 column2" + System.lineSeparator() +
                 "---- -------" + System.lineSeparator() +
-                "one  1      " + System.lineSeparator() +
-                "two  2      " + System.lineSeparator() +
+                "one        1" + System.lineSeparator() +
+                "two        2" + System.lineSeparator() +
                 "..." + System.lineSeparator() +
                 "4 rows x 2 columns", p.toString(df));
     }
@@ -55,10 +55,10 @@ public class TabularPrinter_DataFrameTest {
         assertEquals(System.lineSeparator() +
                 "col1 c..2" + System.lineSeparator() +
                 "---- ----" + System.lineSeparator() +
-                "one  1   " + System.lineSeparator() +
-                "two  2   " + System.lineSeparator() +
-                "t..e 3   " + System.lineSeparator() +
-                "four 4   " + System.lineSeparator() +
+                "one     1" + System.lineSeparator() +
+                "two     2" + System.lineSeparator() +
+                "t..e    3" + System.lineSeparator() +
+                "four   44" + System.lineSeparator() +
                 "4 rows x 2 columns", p.toString(df));
     }
 }