+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #21 Add LocalDate and LocalDateTime converters 
 * #23 Simplifying join API
+* #24 Tabular printer should right-align numbers and booleans
 * #25 CsvLoader to provide shortcuts for datetime and number conversions
 * #26 IntSeries / IntMutableList for joins and filters
 * #27 IntSeries / IntMutableList for sort and group